back/requirements.txt
torch
diffusers
transformers
numpy
pillow
flask
accelerate
xformers
python-dotenv
